requirements.txt
appnope==0.1.0
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==19.0.2
scipy==1.5.2
six==1.15.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1